delete dialog implemented

+++ frontend/src/app/blogsummary/blogsummary.component.ts
@@ -3,6 +3,7 @@ import { BlogService } from '../blog.service';
 import { Blog } from '../../../../api/grpc-web-ts/blogposts_pb';
 import { BlogdialogComponent } from '../blogdialog/blogdialog.component';
 import { MatDialog, MatDialogRef } from '@angular/material';
+import { DeletedialogComponent } from '../deletedialog/deletedialog.component';
 
 @Component({
   selector: 'app-blogsummary',
@@ -11,7 +12,8 @@ import { MatDialog, MatDialogRef } from '@angular/material';
 })
 export class BlogsummaryComponent implements OnInit {
   private blogs: Blog[];
-  private dialogRef: MatDialogRef<BlogdialogComponent>;
+  private blogDialog: MatDialogRef<BlogdialogComponent>;
+  private deleteDialog: MatDialogRef<DeletedialogComponent>;
 
   constructor(private blogService: BlogService, public  dialog: MatDialog) {
     this.blogService.getBlogs((posts: Blog[]) => this.assignBlogs(posts));
@@ -30,32 +32,38 @@ export class BlogsummaryComponent implements OnInit {
   }
 
   createDialog() {
-    if (this.dialogRef) {
-      this.dialogRef.close();
+    if (this.blogDialog) {
+      this.blogDialog.close();
     }
-    this.dialogRef = this.dialog.open(BlogdialogComponent, {
+    this.blogDialog = this.dialog.open(BlogdialogComponent, {
       width: '80%',
     });
   }
 
   createBlog() {
     this.createDialog();
-    this.dialogRef.componentInstance.setFormTitle('Create Blog');
-    this.dialogRef.afterClosed().subscribe(result => {
+    this.blogDialog.componentInstance.setFormTitle('Create Blog');
+    this.blogDialog.afterClosed().subscribe(result => {
       this.blogService.createBlog(result);
     });
   }
 
   editBlog() {
     this.createDialog();
-    this.dialogRef.componentInstance.setFormTitle('Edit Blog');
-    this.dialogRef.componentInstance.setEditValues(this.blogs[1]);
-    this.dialogRef.afterClosed().subscribe(result => {
-      this.blogService.editBlog(result);
+    this.blogDialog.componentInstance.setFormTitle('Edit Blog');
+    this.blogDialog.componentInstance.setEditValues(this.blogs[1]);
+    this.blogDialog.afterClosed().subscribe(result => {
+      if (result) { this.blogService.editBlog(result); }
     });
   }
 
   deleteBlog(blogId: number) {
-    this.blogService.deleteBlog(blogId);
+    if (this.deleteDialog) {
+      this.deleteDialog.close();
+    }
+    this.deleteDialog = this.dialog.open(DeletedialogComponent, {});
+    this.deleteDialog.afterClosed().subscribe(result => {
+      if (result) { this.blogService.deleteBlog(blogId); }
+    });
   }
 }
